reinsert byte casts

+++ trunk/bravo/io/Pathing.java
@@ -541,7 +541,7 @@ represents the distination square of a move.
 //Find the shortest path among all paths. Return null if no valid paths available
     public static void main(String[] args)
     {
-        Move m=new Move(0x40,0x22);
+        Move m=new Move((byte)0x40,(byte)0x22);
         Pathing p=new Pathing();
         int[] a=p.path(m);
         for(int i=0;i<a.length;i++)
